chore: move the lens dev page to a .svelte file
